+++ benches/symbolize.rs
@@ -75,7 +75,7 @@ fn symbolize_dwarf_no_lines() {
 
     let result = results.first().unwrap();
     assert_eq!(result.name, "abort_creds");
-    assert_eq!(result.line, 0);
+    assert_eq!(result.line, None);
 }
 
 /// Symbolize an address in a DWARF file, end-to-end, i.e., including all
@@ -97,7 +97,7 @@ fn symbolize_dwarf() {
 
     let result = results.first().unwrap();
     assert_eq!(result.name, "abort_creds");
-    assert_eq!(result.line, 534);
+    assert_eq!(result.line, Some(534));
 }
 
 /// Symbolize an address in a GSYM file, end-to-end, i.e., including all