+++ examples/addr2ln_pid.rs
@@ -48,19 +48,18 @@ print its symbol, the file name of the source, and the line number.",
                 let Sym {
                     name,
                     addr: sym_addr,
+                    offset,
                     path,
                     line,
                     ..
                 } = sym;
 
                 println!(
-                    "{addr_fmt} {name} @ 0x{sym_addr:x}+0x{:x} {}:{line}",
-                    addr - sym_addr,
+                    "{addr_fmt} {name} @ 0x{sym_addr:x}+0x{offset:x} {}:{line}",
                     path.display(),
                 );
             }
         }
     }
-
     Ok(())
 }