+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Renamed `symbolize::SymbolizedResult` to `Sym` and made it
   non-exhaustive
+  - Changed `Sym::line` to be of type `u32` and `Sym::column` to `u16`
 - Added additional end-to-end benchmarks
   - Added benchmark result summary to CI runs
 - Fixed spurious maps file path creation for low addresses as part of